+++ network.py
@@ -20,15 +20,16 @@ def send_message(IP, PORT, MESSAGE):
         sock.close()
 
 def start_listener(ip, port, callback, stop):
+    # Stop = amount of time script will loop until terminating
     import socket
     sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
     sock.bind((ip, port))
     print(f"Listening on {ip}:{port}...")
-    while True:
+    i = 0
+    while i < stop:
         data, addr = sock.recvfrom(1024)
         callback(data.decode(), addr)
-        if stop:
-            break
+        i += 1
 
 def ping(ip, port, timeout=2):
     sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)